Validation and parsing methods improved
When invalid numbers were entered in DateField components,
validation function was taking too long. So Date.isJalaliValid
method improved.

Also this problem was happening in Date.parseJalali method. In
addition, if we would enter just a number in a DateField, it was
using Gregorian calendar for calculations. This was because of
altFormats config of DateField. A few other formats added to
DateField for better handling of this condition. Also these formats
added to Date.parseFunctions.